requirements.txt
streamlit
pillow
pytesseract
matplotlib
numpy
opencv-python
